#6 feat: 네이버 지도 컴포넌트 분리
